move unsaved check up to +page file
