+++ web-local/src/routes/(application)/model/[name]/+page.svelte
@@ -4,23 +4,22 @@
   import { EntityType } from "@rilldata/web-common/features/entity-management/types";
   import { featureFlags } from "@rilldata/web-common/features/feature-flags";
   import { ModelWorkspace } from "@rilldata/web-common/features/models";
+  import UnsavedSourceDialog from "@rilldata/web-common/features/sources/editor/UnsavedSourceDialog.svelte";
   import { createRuntimeServiceGetFile } from "@rilldata/web-common/runtime-client";
   import { runtime } from "@rilldata/web-common/runtime-client/runtime-store";
   import { error } from "@sveltejs/kit";
   import { onMount } from "svelte";
+  import { goto, beforeNavigate } from "$app/navigation";
+  import { hasUnsavedChanges } from "@rilldata/web-common/features/models/workspace/Editor.svelte";
 
   const { readOnly } = featureFlags;
+
+  let interceptedUrl: string | null = null;
 
   $: modelName = $page.params.name;
 
   /** If ?focus, tell the page to focus the editor as soon as available */
   $: focusEditor = $page.url.searchParams.get("focus") === "";
-
-  onMount(() => {
-    if ($readOnly) {
-      throw error(404, "Page not found");
-    }
-  });
 
   $: fileQuery = createRuntimeServiceGetFile(
     $runtime.instanceId,
@@ -38,6 +37,34 @@
       },
     },
   );
+
+  onMount(() => {
+    if ($readOnly) {
+      throw error(404, "Page not found");
+    }
+  });
+
+  beforeNavigate((e) => {
+    if (!$hasUnsavedChanges || interceptedUrl) return;
+
+    e.cancel();
+
+    if (e.to) {
+      interceptedUrl = e.to.url.href;
+    }
+  });
+
+  async function handleConfirm() {
+    if (interceptedUrl) {
+      await goto(interceptedUrl);
+    }
+
+    interceptedUrl = null;
+  }
+
+  function handleCancel() {
+    interceptedUrl = null;
+  }
 </script>
 
 <svelte:head>
@@ -46,4 +73,12 @@
 
 {#if $fileQuery.data}
   <ModelWorkspace {modelName} focusEditorOnMount={focusEditor} />
+{/if}
+
+{#if interceptedUrl}
+  <UnsavedSourceDialog
+    context="model"
+    on:confirm={handleConfirm}
+    on:cancel={handleCancel}
+  />
 {/if}
